backend/requirements.txt

uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic[email]==2.7.1
pyjwt==2.8.0
bcrypt==4.1.3
python-multipart==0.0.9
# psycopg2-binary==2.9.9
